CompareOptions yields cmp.Options
