+++ tests/templates/deployment/expected.yaml
@@ -6,7 +6,7 @@ metadata:
   labels:
     app: hello-again
 spec:
-  # replicas: 3
+  replicas: 3
   selector:
     matchLabels:
       app: hello-again
