+++ tests/templates/deployment/template.yaml
@@ -12,7 +12,7 @@ spec:
       labels:
         app: $(workload.metadata.name)$
     spec:
-      # replicas: 3
+      replicas: 3
       selector:
         matchLabels:
           app: $(workload.metadata.name)$
